🎨 cmd_parser::models
